Add Customers and Suppliers

+++ src/main/java/com/microservices/users_services/services/CustomerService.java
@@ -3,7 +3,6 @@ package com.microservices.users_services.services;
 import java.util.List;
 import java.util.Optional;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.microservices.users_services.models.Customer;
@@ -11,10 +10,14 @@ import com.microservices.users_services.repositories.CustomerRepository;
 
 @Service
 public class CustomerService {
-    @Autowired
+    //@Autowired
     private CustomerRepository customerRepository;
 
-    public Customer create(Customer customer) {
+    public CustomerService(CustomerRepository customerRepository) {
+        this.customerRepository = customerRepository;
+    }
+
+    public Customer createCustomer(Customer customer) {
         return customerRepository.save(customer);
     }
 
@@ -22,7 +25,7 @@ public class CustomerService {
         return customerRepository.findById(id);
     }
 
-    public List<Customer> getAllCustomers() {
+    public List<Customer> getCustomers() {
         return customerRepository.findAll();
     }
 
@@ -31,7 +34,25 @@ public class CustomerService {
         return customerRepository.save(customer);
     }
 
-    public void delete(String id) {
-        customerRepository.deleteById(id);
+    public Customer updateCustomer(String id, Customer updateCustomer) {
+        return customerRepository.findById(id).map(customer -> {
+            if (updateCustomer.getName() != null)
+                customer.setName(updateCustomer.getName());
+            if (updateCustomer.getEmail() != null)
+                customer.setEmail(updateCustomer.getEmail());
+            if (updateCustomer.getPhone() != null)
+                customer.setPhone(updateCustomer.getPhone());
+
+            return customerRepository.save(customer);
+        }).orElseThrow(() -> new IllegalArgumentException("Customer not found with id: " + id));
+    }
+
+    public Boolean deleteCustomer(String id) {
+        if (customerRepository.existsById(id)) {
+            customerRepository.deleteById(id);
+            return true;
+        } else {
+            throw new IllegalArgumentException("Customer not found with id: " + id);
+        }
     }
 }
